LANGCHAIN/requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
bs4
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain-huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq
arxiv
wikipedia
duckduckgo-search
mysql-connector-python
SQLAlchemy
langchain-cli
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface-hub
gradio
